+++ commands/export.go
@@ -6,22 +6,22 @@ import (
 	"os"
 
 	"github.com/transcriptaze/midiasm/midi"
-	"github.com/transcriptaze/midiasm/ops/export"
+	impl "github.com/transcriptaze/midiasm/ops/export"
 )
 
-type Export struct {
+type export struct {
 	out string
 }
 
-var EXPORT = Export{}
+var Export = export{}
 
-func (x *Export) Flagset(flagset *flag.FlagSet) *flag.FlagSet {
+func (x *export) Flagset(flagset *flag.FlagSet) *flag.FlagSet {
 	flagset.StringVar(&x.out, "out", "", "Output file path (or directory for split files)")
 
 	return flagset
 }
 
-func (x Export) Help() {
+func (x export) Help() {
 	fmt.Println()
 	fmt.Println("  Extracts the MIDI information as JSON for use with other tools (e.g. jq).")
 	fmt.Println()
@@ -42,7 +42,7 @@ func (x Export) Help() {
 	fmt.Println()
 }
 
-func (x Export) Execute(flagset *flag.FlagSet) error {
+func (x export) Execute(flagset *flag.FlagSet) error {
 	filename := flagset.Arg(0)
 
 	smf, err := decode(filename)
@@ -62,8 +62,8 @@ func (x Export) Execute(flagset *flag.FlagSet) error {
 	return x.execute(smf)
 }
 
-func (x Export) execute(smf *midi.SMF) error {
-	op, err := export.NewExport()
+func (x export) execute(smf *midi.SMF) error {
+	op, err := impl.NewExport()
 	if err != nil {
 		return err
 	}
@@ -71,7 +71,7 @@ func (x Export) execute(smf *midi.SMF) error {
 	return x.write(op, smf)
 }
 
-func (x Export) write(op *export.Export, smf *midi.SMF) error {
+func (x export) write(op *impl.Export, smf *midi.SMF) error {
 	out := os.Stdout
 
 	if x.out != "" {