+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## [Unreleased]
 
 
+### Updated
+1. Reworked TSV plugin as a builtin command.
+
+
 ## [0.2.0](https://github.com/transcriptaze/midiasm/releases/tag/v0.2.0) - 2024-05-12
 
 ### Added